+++ lib/ooxml_parser/common_parser/common_data/table/table_properties/table_borders.rb
@@ -1,14 +1,18 @@
 module OoxmlParser
   # Class for describing Table Border Propertie
   class TableBorders < OOXMLDocumentObject
-    # @return [BordersProperties] left border propertie
+    # @return [BordersProperties] left border property
     attr_accessor :left
-    # @return [BordersProperties] right border propertie
+    # @return [BordersProperties] right border property
     attr_accessor :right
-    # @return [BordersProperties] top border propertie
+    # @return [BordersProperties] top border property
     attr_accessor :top
-    # @return [BordersProperties] bottom border propertie
+    # @return [BordersProperties] bottom border property
     attr_accessor :bottom
+    # @return [BordersProperties] inside vertical property
+    attr_accessor :inside_vertical
+    # @return [BordersProperties] inside horizontal property
+    attr_accessor :inside_horizontal
 
     # Parse Table Borders data
     # @param [Nokogiri::XML:Element] node with Table Borders data
@@ -25,6 +29,10 @@ module OoxmlParser
           borders.right = BordersProperties.parse(cell_borders_node)
         when 'bottom'
           borders.bottom = BordersProperties.parse(cell_borders_node)
+        when 'insideV'
+          borders.inside_vertical = BordersProperties.parse(cell_borders_node)
+        when 'insideH'
+          borders.inside_horizontal = BordersProperties.parse(cell_borders_node)
         end
       end
       borders